refactor: cross entropy trainer
