+++ trackers/core/reid/callbacks.py
@@ -69,7 +69,7 @@ class TensorboardCallback(BaseCallback):
 
 
 class WandbCallback(BaseCallback):
-    def __init__(self, config: dict[str, Any]) -> None:
+    def __init__(self, config: Optional[dict[str, Any]] = {}) -> None:
         import wandb
 
         self.run = wandb.init(config=config) if not wandb.run else wandb.run  # type: ignore
